+++ simple_live_app/lib/modules/settings/danmu_settings_page.dart
@@ -173,7 +173,7 @@ class DanmuSettingsView extends GetView<AppSettingsController> {
                   onChanged: (e) {
                     controller.setDanmuSpeed(e.toDouble());
                     updateDanmuOption(
-                      danmakuController?.option.copyWith(duration: e.toInt()),
+                      danmakuController?.option.copyWith(duration: e.toDouble()),
                     );
                   },
                 ),
